basic extension completed and tested
