+++ extension/extension.cmake
@@ -1,5 +1,7 @@
 add_library(extension INTERFACE)
 
+target_sources(extension INTERFACE ${CMAKE_CURRENT_LIST_DIR}/extension.cpp)
+
 target_include_directories(extension INTERFACE ${CMAKE_CURRENT_LIST_DIR})
 
 target_link_libraries(extension INTERFACE motor2040 button pid plasma pwm_cluster pwm motor_cluster motor)
